requirements.txt
python-dotenv
hubspot-api-client
psycopg2-binary pandas
